주석 설명 추가
