--- src/main/java/com/banghyang/auth/domain/OauthMemberRoleType.java
@@ -1,6 +1,0 @@
-package com.banghyang.auth.domain;
-
-public enum OauthMemberRoleType {
-    USER,
-    ADMIN
-}
