--- src/main/java/com/banghyang/auth/infra/oauth/config/HttpInterfaceConfig.java
@@ -1,24 +1,0 @@
-package com.banghyang.auth.infra.oauth.config;
-
-import com.banghyang.auth.infra.oauth.kakao.client.KakaoApiClient;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.reactive.function.client.WebClient;
-import org.springframework.web.reactive.function.client.support.WebClientAdapter;
-import org.springframework.web.service.invoker.HttpServiceProxyFactory;
-
-@Configuration
-public class HttpInterfaceConfig {
-    // Http Interface Client 구현체를 빈으로 등록
-    @Bean
-    public KakaoApiClient kakaoApiClient() {
-        return createHttpInterface(KakaoApiClient.class);
-    }
-
-    private <T> T createHttpInterface(Class<T> clazz) {
-        WebClient webClient = WebClient.create();
-        HttpServiceProxyFactory build = HttpServiceProxyFactory
-                .builder(WebClientAdapter.forClient(webClient)).build();
-        return build.createClient(clazz);
-    }
-}
