--- src/main/java/com/banghyang/auth/infra/presentation/OauthController.java
@@ -1,38 +1,0 @@
-package com.banghyang.auth.infra.presentation;
-
-import com.banghyang.auth.application.OauthService;
-import com.banghyang.auth.domain.OauthServerType;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import lombok.SneakyThrows;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-@RequestMapping("/oauth")
-@RestController
-@RequiredArgsConstructor
-public class OauthController {
-
-    private final OauthService oauthService;
-
-    @SneakyThrows
-    @GetMapping("/{oauthServerType}")
-    ResponseEntity<Void> redirectAuthCodeRequestUrl(
-            @PathVariable OauthServerType oauthServerType,
-            HttpServletResponse response
-    ) {
-        String redirectUrl = oauthService.getAuthCodeRequestUrl(oauthServerType);
-        response.sendRedirect(redirectUrl);
-        return ResponseEntity.ok().build();
-    }
-
-    @GetMapping("/login/{oauthServerType}")
-    ResponseEntity<Long> login(
-            @PathVariable OauthServerType oauthServerType,
-            @RequestParam("code") String code
-    ) {
-        Long login = oauthService.login(oauthServerType, code);
-
-        return ResponseEntity.ok(login);
-    }
-}
