--- src/main/java/com/banghyang/auth/infra/presentation/OauthServerTypeConverter.java
@@ -1,11 +1,0 @@
-package com.banghyang.auth.infra.presentation;
-
-import com.banghyang.auth.domain.OauthServerType;
-import org.springframework.core.convert.converter.Converter;
-
-public class OauthServerTypeConverter implements Converter<String, OauthServerType> {
-    @Override
-    public OauthServerType convert(String source) {
-        return OauthServerType.fromName(source);
-    }
-}
