+++ src/main/java/com/banghyang/oauth/domain/OauthId.java
@@ -1,0 +1,23 @@
+package com.banghyang.oauth.domain;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Embeddable;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Embeddable
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+@Getter
+public class OauthId {
+    @Column(nullable = false, name = "oauth_server_id")
+    private String oauthServerId;
+
+    @Enumerated(EnumType.STRING)
+    @Column(nullable = false, name = "oauth_server")
+    private OauthServerType oauthServerType;
+}
